Make tests pass on rakudo 2018.05
Hashes are unordered, so expecting a specific order in tests is
wrong. In the past Rakudo used to give out results in exactly the same
order every run (which is why these tests were working), but now it is
randomized due to security reasons.

Perhaps this module should sort the elements in refine-str, but that's
not for me to decide. For now, let's at least make these tests pass.